+++ backend/controllers/bills-controller.js
@@ -44,7 +44,7 @@ function createBills(req, res, next) {
 
   newValuesStr = 'values ';
   req.body.map((a, i) => {
-    newValuesStr = newValuesStr + '(' + parseInt(a.amount) + ',' + parseInt(a.categoryid) + ',' + parseInt(a.userid) + ',\'' +  a.date + '\')';
+    newValuesStr = newValuesStr + '(' + parseFloat(a.amount) + ',' + parseInt(a.categoryid) + ',' + parseInt(a.userid) + ',\'' +  a.date + '\')';
     if (i+1 != req.body.length) {
       newValuesStr = newValuesStr + ',';
     }
@@ -78,7 +78,7 @@ function updateBills(req, res, next) {
 
   oldValuesStr = 'values ';
   req.body.map((a, i) => {
-    oldValuesStr = oldValuesStr + '(' + parseInt(a.id) + ',' + parseInt(a.userid) + ',' + parseInt(a.categoryid) + ',' + parseInt(a.amount) + ',\'' +  a.date + '\'::date)';
+    oldValuesStr = oldValuesStr + '(' + parseInt(a.id) + ',' + parseInt(a.userid) + ',' + parseInt(a.categoryid) + ',' + parseFloat(a.amount) + ',\'' +  a.date + '\'::date)';
     if (i+1 != req.body.length) {
       oldValuesStr = oldValuesStr + ',';
     }
